Move NavItems into shared models
