+++ src/Portfolio/Layout/Components/Navbar.razor.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Components;
 using Portfolio.Navigation;
+using Models;
 
 namespace Portfolio.Layout.Components;
